+++ lib/src/rules/max_length_rule.dart
@@ -1,23 +1,23 @@
 import 'package:flrx_validator/src/rules/rule.dart';
 
 /// A [Rule] subclass validating if the length of input is less than [maxLength].
-class MaxLengthRule extends Rule<String> {
+class MaxLengthRule<T extends dynamic> extends Rule<T> {
   final int maxLength;
 
   MaxLengthRule(this.maxLength, {String validationMessage})
-      : super(validationMessage);
+      : assert(T == List || T == Map || T == String || T == dynamic),
+        super(validationMessage);
 
   @override
-  String onValidate(String entityName, String value) {
+  String onValidate(String entityName, T value) {
     if (value.length <= maxLength) {
       return null;
     }
-    return validationMessage ??
-        ":entity should be less than :maxLength characters";
+    return validationMessage ?? ':entity length should be less than :maxLength';
   }
 
   @override
   Map<String, String> getRuleSpecificParams() {
-    return <String, String>{"maxLength": maxLength.toString()};
+    return {'maxLength': maxLength.toString()};
   }
 }