+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+## [v0.5.0]
+### Added
+Added `NumberRule` which checks if the value is a valid number or not
+Added `MinRule` which checks if the value is greater a specific number or not
+Added `MaxRule` which checks if the value is less than a specific number or not
+
+### Changed
+`MinLengthRule`, `MaxLengthRule` now supports List and Map
+`RequiredRule` now supports List, Map and other Data types. In case of other Data type, nullability of the value is tested
+Update Project Style
+Migrate Example to AndroidX
+Update Docs Dependencies
+
+### Removed
+- `Validator.build()` deprecated in [v0.4.0+2]
+
 ## [v0.4.0+2]
 
 ### Added
@@ -38,6 +54,7 @@
 
 Initial Release
 
+[v0.5.0]: https://github.com/flrx/validator/compare/v0.5.0...v0.4.0+2
 [v0.4.0+2]: https://github.com/flrx/validator/compare/v0.4.0+2...v0.3.0
 [v0.3.0]: https://github.com/flrx/validator/compare/v0.3.0...v0.2.0
 [v0.2.0]: https://github.com/flrx/validator/compare/v0.2.0...v0.1.0